chat/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiodns==3.0.0
aiohttp[speedups]==3.7.4.post0
alembic==1.6.5
aniso8601==7.0.0
asgiref==3.4.1; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
asyncpg==0.23.0
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.6
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1; python_version >= '3.6'
dnspython==1.16.0; python_version >= '3.6'
email-validator==1.1.3
fastapi==0.66.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0; python_version >= '3'
h11==0.12.0; python_version >= '3.6'
httptools==0.2.0
idna==3.2; python_full_version >= '3.5.0'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
motor==2.4.0
multidict==5.1.0; python_version >= '3.6'
pipfile==0.0.2
promise==2.3
psycopg2-binary==2.9.1
pycares==4.0.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic[email]==1.8.2
pyjwt==2.1.0
pymongo[srv]==3.12.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.18.0
python-editor==1.0.4
pyyaml==5.4.1
rx==1.6.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.21
starlette==0.14.2; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
uvicorn[standard]==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
yarl==1.6.3; python_version >= '3.6'
